Reorganize components folder structure
