+++ s3-react-basics/src/components/Expenses/ExpenseItem.js
@@ -1,5 +1,5 @@
 import "./ExpenseItem.css";
-import ExpenseDate from "../ExpenseDate/ExpenseDate";
+import ExpenseDate from "./ExpenseDate";
 import Card from "../UI/Card";
 
 function ExpenseItem({ title, date, amount }) {
